resolves #58: Ship vendored fontawesome for icons
We would like antora-ui to include fontawesome so that :icons:font
from asciidoc can work and we can have fun icons in documents.

When :icons:font is set, asciidoctor also unconditionally makes HTML
for admonitions with icons that makes the admonition title like

<div class="admonitionblock tip">
 <table>
  <tbody>
   <tr>
    <td class="icon">
      <i class="fa icon-tip" title="Tip"></i>
    </td>
   </tr>
  </tbody>
 </table>
</div>

It's not an option to turn this off in asciidoctor.  So we need CSS to
deal with this when :icons:font is on so that admonitions still appear
correct.  The approach is to add the icon as a ::before pseudo-element
in the FontAwesome font, and the title text as an ::after element
styled in the default font.  With some mapping of the icon to a
suitable unicode FA icon, we then get the admonition title with a nice
looking icon and text next to it.

The other point of contention is ensuring that we ship FA locally to
avoid hitting a CDN or similar.  We copy in the fonts and main CSS as
vendored files from the official npm distribution via gulp.  We make
sure the "Font Awesome 7 Free" family finds the fonts from our local
copy shipped in the static directory.  This way the static site is
self-contained.

For history, this is a refinement of

  https://gitlab.com/antora/antora-ui-default/-/merge_requests/184

which was a refinement of

  https://gitlab.com/antora/antora-ui-default/-/merge_requests/121

The colors are from

  https://github.com/mojavelinux/asciidoctor-docs-ui/commit/4eba6c71c2ce31488324544564710526d1784bb9